Tons of updates

+++ src/main/java/com/ivolodin/entities/StationConnect.java
@@ -21,6 +21,6 @@ public class StationConnect {
     private Station to;
 
     @Column
-    private long distanceInMinutes;
+    private Integer distanceInMinutes;
 
 }
